commit 5169bd6303b17f9744bfdacfbbe473c945e7dd3f
Author: Rogerio Oliveira <roliveira@cyclomedia.com>
Date:   Tue Apr 4 13:24:55 2023 +0200

    added .vscode directory into gitignore;
    added some formating;
    refactored the auth method;
    - handle-errors
    added a new return error condition when we have a missing header;

commit 1316de382c3dee28c1b0156e6e5aa582b1fcfcc7
Author: Rogerio Oliveira <roliveira@cyclomedia.com>
Date:   Tue Apr 4 00:27:36 2023 +0200

    - handle-errors
    added sqlx postgres to use the error list;
    added argon2 to use hash password error list;
    added custom errors for wrong password, when it's not possible to verify password, database queries, decrypt token, when unauthorized;
    - webapp
    added missing table creation;
    added new columns and account table;
    added account type;
    added authentication route;
    added authentication methods on store;
    changed answer and question routes to work with account sessions;
    added authorization to protect sensible routes;

+++ src/routes/question.rs
@@ -2,15 +2,18 @@
 
 use std::collections::HashMap;
 
+use handle_errors::Error;
 use tracing::{event, instrument, Level};
 use warp::hyper::StatusCode;
 
 use crate::{
+    services::profanity::check_profanity,
     store::Store,
     types::{
+        account::Session,
         pagination::{extract_pagination, Pagination},
         question::{NewQuestion, Question},
-    }, services::profanity::check_profanity,
+    },
 };
 
 #[instrument]
@@ -36,19 +39,21 @@ pub async fn get_questions(
 }
 
 pub async fn add_question(
+    session: Session,
     store: Store,
     new_question: NewQuestion,
 ) -> Result<impl warp::Reply, warp::Rejection> {
+    let account_id = session.account_id;
 
     let title = check_profanity(new_question.title);
     let content = check_profanity(new_question.content);
- 
+
     let (title, content) = tokio::join!(title, content);
- 
+
     if title.is_err() {
         return Err(warp::reject::custom(title.unwrap_err()));
     }
- 
+
     if content.is_err() {
         return Err(warp::reject::custom(content.unwrap_err()));
     }
@@ -59,7 +64,7 @@ pub async fn add_question(
         tags: new_question.tags,
     };
 
-    match store.add_question(question).await {
+    match store.add_question(question, account_id).await {
         Ok(question) => Ok(warp::reply::json(&question)),
         Err(e) => Err(warp::reject::custom(e)),
     }
@@ -68,42 +73,57 @@ pub async fn add_question(
 // pay attention!!! the signature need to follow this order!!! param, store, item to be updated
 pub async fn update_question(
     id: i32,
+    session: Session,
     store: Store,
     question: Question,
 ) -> Result<impl warp::Reply, warp::Rejection> {
+    let account_id = session.account_id;
 
-    let title = check_profanity(question.title);
-    let content = check_profanity(question.content);
- 
-    let (title, content) = tokio::join!(title, content);
- 
-    if title.is_err() {
-        return Err(warp::reject::custom(title.unwrap_err()));
-    }
- 
-    if content.is_err() {
-        return Err(warp::reject::custom(content.unwrap_err()));
-    }
- 
-    let question = Question {
-        id: question.id,
-        title: title.unwrap(),
-        content: content.unwrap(),
-        tags: question.tags,
-    };
-        
-    match store.update_question(question, id).await {
-        Ok(res) => Ok(warp::reply::json(&res)),
-        Err(e) => Err(warp::reject::custom(e)),
+    if store.is_question_owner(id, &account_id).await? {
+        let title = check_profanity(question.title);
+        let content = check_profanity(question.content);
+
+        let (title, content) = tokio::join!(title, content);
+
+        if title.is_err() {
+            return Err(warp::reject::custom(title.unwrap_err()));
+        }
+
+        if content.is_err() {
+            return Err(warp::reject::custom(content.unwrap_err()));
+        }
+
+        let question = Question {
+            id: question.id,
+            title: title.unwrap(),
+            content: content.unwrap(),
+            tags: question.tags,
+        };
+
+        match store.update_question(question, id, account_id).await {
+            Ok(res) => Ok(warp::reply::json(&res)),
+            Err(e) => Err(warp::reject::custom(e)),
+        }
+    } else {
+        Err(warp::reject::custom(Error::Unauthorized))
     }
 }
 
-pub async fn delete_question(id: i32, store: Store) -> Result<impl warp::Reply, warp::Rejection> {
-    match store.delete_question(id).await {
-        Ok(_) => Ok(warp::reply::with_status(
-            format!("Question {} deleted", id),
-            StatusCode::OK,
-        )),
-        Err(e) => Err(warp::reject::custom(e)),
+pub async fn delete_question(
+    id: i32,
+    session: Session,
+    store: Store,
+) -> Result<impl warp::Reply, warp::Rejection> {
+    let account_id = session.account_id;
+    if store.is_question_owner(id, &account_id).await? {
+        match store.delete_question(id).await {
+            Ok(_) => Ok(warp::reply::with_status(
+                format!("Question {} deleted", id),
+                StatusCode::OK,
+            )),
+            Err(e) => Err(warp::reject::custom(e)),
+        }
+    } else {
+        Err(warp::reject::custom(Error::Unauthorized))
     }
 }
